testing/web-platform/tests/html/editing/dnd/the-draggable-attribute/draggable_attribute.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset='utf-8'>
    <title>HTML Test: draggable_attribute</title>
    <link rel='author' title='Intel' href='http://www.intel.com'>
    <link rel='help' href='http://www.w3.org/html/wg/drafts/html/CR/editing.html#the-draggable-attribute'>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
  </head>

  <body>
    <div id='log'> </div>

    <script>
      //The "data_all" below has been stolen from @Ms2ger's html/semantics/interfaces.html, thanks!
      var data_all = [
        ["abbr", ""],
        ["acronym", ""],
        ["address", ""],
        ["applet", "Applet"],
        ["audio", "Audio"],
        ["b", ""],
        ["base", "Base"],
        ["basefont", "BaseFont"],
        ["bdo", ""],
        ["bgsound", "Unknown"],
        ["big", ""],
        ["blink", "Unknown"],
        ["blockquote", "Quote"],
        ["body", "Body"],
        ["br", "BR"],
        ["button", "Button"],
        ["canvas", "Canvas"],
        ["caption", "TableCaption"],
        ["center", ""],
        ["cite", ""],
        ["code", ""],
        ["col", "TableCol"],
        ["colgroup", "TableCol"],
        ["dd", ""],
        ["del", "Mod"],
        ["dfn", ""],
        ["dir", "Directory"],
        ["div", "Div"],
        ["dl", "DList"],
        ["dt", ""],
        ["em", ""],
        ["embed", "Embed"],
        ["fieldset", "FieldSet"],
        ["font", "Font"],
        ["form", "Form"],
        ["frame", "Frame"],
        ["frameset", "FrameSet"],
        ["h1", "Heading"],
        ["h2", "Heading"],
        ["h3", "Heading"],
        ["h4", "Heading"],
        ["h5", "Heading"],
        ["h6", "Heading"],
        ["head", "Head"],
        ["hr", "HR"],
        ["html", "Html"],
        ["i", ""],
        ["iframe", "IFrame"],
        ["image", "Unknown"],
        ["img", "Image"],
        ["input", "Input"],
        ["ins", "Mod"],
        ["isindex", "Unknown"],
        ["kbd", ""],
        ["keygen", "Keygen"],
        ["label", "Label"],
        ["legend", "Legend"],
        ["li", "LI"],
        ["link", "Link"],
        ["listing", ""],
        ["map", "Map"],
        ["marquee", "Marquee"],
        ["menu", "Menu"],
        ["meta", "Meta"],
        ["multicol", "Unknown"],
        ["nobr", ""],
        ["noembed", ""],
        ["noframes", ""],
        ["noscript", ""],
        ["object", "Object"],
        ["ol", "OList"],
        ["optgroup", "OptGroup"],
        ["option", "Option"],
        ["p", "Paragraph"],
        ["param", "Param"],
        ["plaintext", ""],
        ["pre", "Pre"],
        ["q", "Quote"],
        ["s", ""],
        ["samp", ""],
        ["script", "Script"],
        ["select", "Select"],
        ["small", ""],
        ["source", "Source"],
        ["spacer", "Unknown"],
        ["span", "Span"],
        ["strike", ""],
        ["strong", ""],
        ["style", "Style"],
        ["sub", ""],
        ["sup", ""],
        ["table", "Table"],
        ["tbody", "TableSection"],
        ["td", "TableDataCell"],
        ["textarea", "TextArea"],
        ["tfoot", "TableSection"],
        ["th", "TableHeaderCell"],
        ["thead", "TableSection"],
        ["title", "Title"],
        ["tr", "TableRow"],
        ["tt", ""],
        ["u", ""],
        ["ul", "UList"],
        ["var", ""],
        ["video", "Video"],
        ["wbr", ""],
        ["xmp", ""],
        ["section", ""],
        ["nav", ""],
        ["article", ""],
        ["aside", ""],
        ["hgroup", ""],
        ["header", ""],
        ["footer", ""],
        ["time", "Unknown"],
        ["mark", ""],
        ["progress", "Progress"],
        ["meter", "Meter"],
        ["ruby", ""],
        ["rt", ""],
        ["rp", ""],
        ["figure", ""],
        ["datalist", "DataList"],
        ["output", "Output"],
        ["details", "Details"],
        ["command", "Command"],
        ["nextid", "Unknown"],
        ["rb", "Unknown"],
        ["bdi", ""],
        ["data", "Data"],
        ["a", "Anchor"],
        ["area", "Area"],
        ["dialog", "Dialog"],
        ["figcaption", ""],
        ["summary", ""],
        ["track", "Track"]
      ];

      data_all.forEach(function(a) {
        test(function() {
          var eElement = document.createElement(a[0]);
          assert_inherits(eElement, 'draggable', 'Element ' + a[0] +' should have draggable property');
        }, 'Element ' + a[0] +' should have draggable property');
      });

      function run_test(element, element_name, exp) {
        if (exp) {
          assert_true(element.draggable, 'Element ' + element_name +' should be draggable');
        } else {
          assert_false(element.draggable, 'Element ' + element_name +' should not be draggable');
        }
      }

      function run_idl_test(element, element_name, exp) {
        if (exp) {
          assert_equals(element.getAttribute('draggable'), 'true', 'Element ' + element_name +' should be draggable');
        } else {
          assert_equals(element.getAttribute('draggable'), 'false', 'Element ' + element_name +' should not be draggable');
        }
      }

      data_all.forEach(function(a) {

        test(function() {
          //Default values for elements
          //If the element is an img element, or, if the element is an a element with an href content attribute,
          //the draggable IDL attribute must return true.
          var eElement = document.createElement(a[0]);
          switch (a[0]) {
            case 'a':
              eElement.setAttribute('href', 'http://w3.org');
              run_test(eElement, 'a', true);
              break;
            case 'img':
              run_test(eElement, 'img', true);
              break;
            default:
              run_test(eElement, a[0], false);
          }

          //If an element's draggable content attribute has the state true,
          //the draggable IDL attribute must return true.
          eElement.setAttribute('draggable', 'true');
          run_test(eElement, a[0] + ' draggable=\'true\'', true);

          //If an element's draggable content attribute has the state false,
          //the draggable IDL attribute must return false.
          eElement.setAttribute('draggable', 'false');
          run_test(eElement, a[0] + ' draggable=\'false\'', false);

          //auto values for elements
          //The element's draggable content attribute has the state auto.
          //If the element is an img element, or, if the element is an a element with an href content attribute,
          //the draggable IDL attribute must return true.
          switch (a[0]) {
            case 'a':
              eElement.setAttribute('href', 'http://w3.org');
              eElement.setAttribute('draggable', 'auto');
              run_test(eElement, 'Element ' + 'a' + ' draggable=\'auto\'', true);
              break;
            case 'img':
              eElement.setAttribute('draggable', 'auto');
              run_test(eElement, 'Element ' + 'img' + ' draggable=\'auto\'', true);
              break;
            default:
              run_test(eElement, 'Element ' + a[0] + ' draggable=\'auto\'', false);
          }

          //Foo values for elements
          //The element's draggable content attribute value is not enumerated (true, false, auto) but unexpected.
          //Fallback to defaults
          switch (a[0]) {
            case 'a':
              eElement.setAttribute('href', 'http://w3.org');
              eElement.setAttribute('draggable', 'foo');
              run_test(eElement, 'Element ' + 'a' + ' draggable=\'foo\'', true);
              break;
            case 'img':
              eElement.setAttribute('draggable', 'foo');
              run_test(eElement, 'Element ' + 'img' + ' draggable=\'foo\'', true);
              break;
            default:
              run_test(eElement, 'Element ' + a[0] + ' draggable=\'foo\'', false);
          }

          //An element with a draggable attribute should also have a title attribute
          //that names the element for the purpose of non-visual interactions.
          eElement.setAttribute('title', 'foo as title value');
          assert_equals(typeof eElement.title, 'string', '<' + a[0] + '> draggable block has title attribute');

          //If the draggable IDL attribute is set to the value false,
          //the draggable content attribute must be set to the literal value false.
          eElement.draggable = false;
          run_idl_test(eElement, a[0] + '.getAttribute(\'draggable\') is \'false\'', false);

          //If the draggable IDL attribute is set to the value true,
          //the draggable content attribute must be set to the literal value true.
          eElement.draggable = true;
          run_idl_test(eElement, a[0] + '.getAttribute(\'draggable\') is \'true\'', true);
          }, 'Element ' + a[0] +' draggable attribute test');

        });
    </script>
  </body>
</html>
